requirements.txt
aiofiles
aiohttp
#asyncio==3.4.3
asyncio
beautifulsoup4
dnspython
ffmpeg-python
future==0.18.3
gitpython
httpx[http2]
heroku3
hachoir
motor==3.3.2
pillow==9.5.0
psutil
py-tgcalls==0.9.7
pykeyboard
pymongo==4.6.3
python-dotenv
pyyaml==6.0.1
requests
SafoneAPI==1.0.65
tgcrypto
spotipy
gTTS
pyfiglet
telegraph
speedtest-cli
unidecode
urllib3
wget
git+https://github.com/KurimuzonAkuma/pyrogram@dev
yt-dlp
git+https://github.com/alexmercerind/youtube-search-python@main
git+https://github.com/joetats/youtube_search@master
